+++ tests/Integration/Services/Main/Service/MainTest.php
@@ -39,6 +39,18 @@ class MainTest extends TestCase
         );
     }
 
+    /**
+     * @covers  Bitrix24\SDK\Services\Main\Service\Main::getAvailableMethods
+     * @testdox test methods list
+     * @return void
+     * @throws \Bitrix24\SDK\Core\Exceptions\BaseException
+     * @throws \Bitrix24\SDK\Core\Exceptions\TransportException
+     */
+    public function testList(): void
+    {
+        $this->assertIsArray($this->mainService->getAvailableMethods()->getResponseData()->getResult()->getResultData());
+    }
+
     public function setUp(): void
     {
         $this->mainService = Fabric::getServiceBuilder()->getMainScope()->main();